+++ CHANGELOG.md
@@ -23,6 +23,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Method `sendWithDetails` of class `XrpClient` is deprecated and will be removed in two release cycles.
   Use `sendXrpWithDetails` instead.
 
+## 5.2.0 - 2020-10-15
+
+- A new method `unauthorizeSendingAccount` is added to `XrpClient` which unauthorizes
+  a sender to send XRP to the specified XRPL account that has enabled Deposit Authorization.
+  (See https://xrpl.org/depositpreauth.html)
+
 - A new method `authorizeSendingAccount` is added to `XrpClient` which authorizes
   senders to send XRP to the specified XRPL account.
   (See https://xrpl.org/depositpreauth.html)